CLEANUP: Cleaned up some varnish files which were causing problems.
Makefile.am now uses the markdown readme file and the licence from the
parent directory. The example VCL now uses a Lite data file which is now
coppied when running run.sh for the first time. The licence headers on
all examples have been changed to be commented out correctly, and the
version test changed to reflect the naming of the method.