BUG: Node.js memory increased due to references not being cleared by the garbage collector.
The Match.close() method releases the C/C++ object and frees any memory,
but as node.js has been told it is not a node.js object and should not
clear it up, the node.js object (which will just be a reference to the
C/C++ object) is not freed. The SWIG interface is now the same as all
other languages in that it declares that the getMatch method returns a
new object that should be cleared. The only difference is that the
destructor does not do anything to the C/C++ object, this is done by the
close method. The usage of Match objects is unchanged. The internal
steps are:
1) close method is called,
2) this releases the C/C++ object and any memory allocated,
3) when the garbage collector runs, any references/strings in node.js
are freed.