BUG: Nginx shared memory locks were not being properly handled.
The cache is now locked until the result of the cache fetch is finished
with. The dataset cleanup now ensures that only one process attempts to
free the dataset, and that it is the last process to exit which does so.

Pre-compiled modules have also been rebuilt with this fix, and the makefile optimised to build the modules without the need to build the entire Nginx binary.