FEAT: 51Degrees PHP common API.
The new 51Degrees PHP common API aims to make switch from Cloud based Device-Detection to the On-Premise version much easier for developers. It allows developers to switch without the need to refactor code by providing a common API between the two solutions connected by an interface which communicates with the core software.