BUG: When parameter matches were used alongside caching, proxied requests would contain null headers.
Only the HTTP header matches were cached, so if there were only matches
on a parameter, there was nothing to cache. However, the node was cached
anyway. This meant that when the node was retrieved from the cache in a
subsequent request, there was a problem when proxying the request to
another server due to a null header being added.

Caching is now keyed on both the HTTP headers, and any parameters that
are to be matched. Adding a node to the cache now occurs after all
matching has been carried out. This ensures all matches are treated in
the same way and dealt with by the cache.