Merged PR 1697: BUILD/BUG: Update Nginx Makefile to use an up to date stable release, and fixed bug in PHP common API.
- BUILD: Updated Nginx Makefile to use an up to date stable release.
- BUG: When using an on-premise provider through the PHP common API, a memory leak could occur depending on the PHP setup.
Using the on-premise provider directly gets the existing provider, but the common API created a new provider each time. This adds an extra overhead to the request and could result in a memory leak depending on the PHP setup.
This has now been changed to access the existing provider in the same was as directly using the on-premise provider.