BUG: Fixed memory leak when reloading the provider with a GNU compiled library.
The `FIFTYONEDEGREES_INTERLOCK_INC` and `FIFTYONE_DEGREES_INTERLOCK_DEC`
macros were not consistent across MSVC and GNU comilers. The
`_InterlockedDecrement` method returns the value after decrementing
(similarly for the increment method). However, the
`__sync_fetch_and_add` method returns the value before decrementing (or
incrementing). This meant that the `inUse` counter may not be evaluated
as 0, and therefore not allow the provider to be freed.
The GNU implementation now uses `__sync_add_and_fetch` instead of
`__sync_fetch_and_fetch` so that the logic of the macros agrees with the
usage.

fixes #46